requirements.txt
aiohttp>=3.9.0
PyJWT>=2.8.0
requests>=2.31.0
